+++ tests/char.test.zsh
@@ -22,6 +22,8 @@ setUp() {
   SPACESHIP_CHAR_PREFIX=""
   SPACESHIP_CHAR_SUFFIX=""
   SPACESHIP_CHAR_SYMBOL="➜ "
+  SPACESHIP_CHAR_SYMBOL_SUCCESS="${SPACESHIP_CHAR_SYMBOL}"
+  SPACESHIP_CHAR_SYMBOL_FAILURE="${SPACESHIP_CHAR_SYMBOL}"
   SPACESHIP_CHAR_COLOR_SUCCESS="green"
   SPACESHIP_CHAR_COLOR_FAILURE="red"
   SPACESHIP_CHAR_COLOR_SECONDARY="yellow"
@@ -48,30 +50,23 @@ tearDown() {
 
 test_char() {
   SPACESHIP_CHAR_COLOR_SUCCESS=blue
+  SPACESHIP_CHAR_SYMBOL_SUCCESS="S "
 
-  local expected="%{%B%F{$SPACESHIP_CHAR_COLOR_SUCCESS}%}➜ %{%b%f%}"
+  local expected="%{%B%F{$SPACESHIP_CHAR_COLOR_SUCCESS}%}$SPACESHIP_CHAR_SYMBOL_SUCCESS%{%b%f%}"
   local actual="$(spaceship::testkit::render_prompt)"
 
   assertEquals "render char" "$expected" "$actual"
 }
 
 test_char_failure() {
-  SPACESHIP_CHAR_COLOR_FAILURE=yello
+  SPACESHIP_CHAR_COLOR_FAILURE=yellow
+  SPACESHIP_CHAR_SYMBOL_FAILURE="F "
 
-  local expected="%{%B%F{$SPACESHIP_CHAR_COLOR_FAILURE}%}➜ %{%b%f%}"
+  local expected="%{%B%F{$SPACESHIP_CHAR_COLOR_FAILURE}%}$SPACESHIP_CHAR_SYMBOL_FAILURE%{%b%f%}"
   command false # this command should exit with non-zero code
   local actual="$(spaceship::testkit::render_prompt)"
 
   assertEquals "render char with failure" "$expected" "$actual"
-}
-
-test_char_symbol() {
-  SPACESHIP_CHAR_SYMBOL='-> '
-
-  local expected="%{%B%F{green}%}$SPACESHIP_CHAR_SYMBOL%{%b%f%}"
-  local actual="$(spaceship::testkit::render_prompt)"
-
-  assertEquals "render char with custom symbol" "$expected" "$actual"
 }
 
 test_char_prefix() {