+++ tests/dart.test.zsh
@@ -57,7 +57,7 @@ test_no_files() {
 }
 
 test_dart_upsearch_file() {
-  FILES=( pubspec.yaml pubspec.yml pubspec.lock )
+  FILES=(pubspec.yaml pubspec.yml pubspec.lock)
   for file in $FILES; do
     touch $file
     local expected="%{%B%}via %{%b%}%{%B%F{$SPACESHIP_DART_COLOR}%}${SPACESHIP_DART_SYMBOL}v$DART_VERSION%{%b%f%}"
@@ -68,7 +68,7 @@ test_dart_upsearch_file() {
 }
 
 test_dart_upsearch_dir() {
-  DIRS=( dart_tool )
+  DIRS=(dart_tool)
   for dir in $DIRS; do
     mkdir $dir
     local expected="%{%B%}via %{%b%}%{%B%F{$SPACESHIP_DART_COLOR}%}${SPACESHIP_DART_SYMBOL}v$DART_VERSION%{%b%f%}"
@@ -93,7 +93,7 @@ test_dart_flutter() {
   # alias to mock dart command bundled with flutter
   alias "dart"="printf 'Dart SDK version: 2.18.1 (stable) (Tue Sep 13 11:42:55 2022 +0200) on \"macos_x64\"'\n"
   local DART_VERSION="2.18.1"
-  file="main.dart" 
+  file="main.dart"
   touch $file
   local expected="%{%B%}via %{%b%}%{%B%F{$SPACESHIP_DART_COLOR}%}${SPACESHIP_DART_SYMBOL}v$DART_VERSION%{%b%f%}"
   local actual="$(spaceship::testkit::render_prompt)"
@@ -105,7 +105,7 @@ test_dart_sdk() {
   # alias to mock dart command installed as dart-sdk
   alias "dart"="printf 'Dart SDK version: 2.19.0-edge.efb509c114dcaf54d0a011f717b48893d71ec9c3 (be) (Thu Sep 29 01:58:56 2022 +0000) on \"macos_x64\"'\n"
   local DART_VERSION="2.19.0"
-  file="main.dart" 
+  file="main.dart"
   touch $file
   local expected="%{%B%}via %{%b%}%{%B%F{$SPACESHIP_DART_COLOR}%}${SPACESHIP_DART_SYMBOL}v$DART_VERSION%{%b%f%}"
   local actual="$(spaceship::testkit::render_prompt)"