+++ tests/dir.test.zsh
@@ -25,8 +25,6 @@ setUp() {
   SPACESHIP_DIR_TRUNC=3
   SPACESHIP_DIR_TRUNC_REPO=true
   SPACESHIP_DIR_COLOR="cyan"
-
-  git config --global init.defaultBranch main
 }
 
 oneTimeTearDown() {
@@ -100,8 +98,8 @@ test_dir_trunc_git() {
   mkdir -p $REPO/dir4/dir5
 
   cd $REPO
-  command git config init.defaultBranch > /dev/null
-  command git init > /dev/null
+  command git config init.defaultBranch >/dev/null
+  command git init >/dev/null
 
   cd $REPO/dir4/dir5
 
@@ -115,6 +113,10 @@ test_dir_trunc_git_submodule() {
   local FOLDER="$SHUNIT_TMPDIR/dir/trunc_git_submodule"
   local REPO="$FOLDER/dir1/dir2/dir3"
   local SUBMODULE="$FOLDER/dir1/dir2/dir4"
+
+  # Allow adding submodules via file transport
+  # See: https://bugs.launchpad.net/ubuntu/+source/git/+bug/1993586
+  command git config --global protocol.file.allow always
 
   mkdir -p $REPO
   mkdir -p $SUBMODULE